backend/requirements.txt
streamlit
flask
flask-cors
transformers
torch
Pillow
